Move all Mod.Cnc features to Mods.Common namespace
