+++ OpenRA.Mods.Cnc/CncLoadScreen.cs
@@ -14,10 +14,9 @@ using System.Diagnostics;
 using System.Drawing;
 using System.IO;
 using OpenRA.Graphics;
-using OpenRA.Mods.Common.LoadScreens;
 using OpenRA.Widgets;
 
-namespace OpenRA.Mods.Cnc
+namespace OpenRA.Mods.Common.LoadScreens
 {
 	public sealed class CncLoadScreen : BlankLoadScreen
 	{
